chore: move auth-router
